Adição de nova migration, alterando valores decimais para double e iniciando a tratativa de itens na loja

+++ Bag-and-Shop-app/Application/DTOs/Bag/BagResponseDTO.cs
@@ -7,7 +7,7 @@ namespace Bag_and_Shop_app.Application.DTOs.Character
     {
 
         public int Id { get; set; }
-        public decimal Gold { get; set; }
+        public double Gold { get; set; }
         public double WeightLimit { get; set; }
         public double CurrentWeight { get; set; }
         public int CharacterId { get; set; }
